file dump

+++ misc_c/gdb_debug/make.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-rm makefile.txt
+if [ -f  "makefile.txt" ]; then
+	rm makefile.txt
+fi
+
 echo "all: prog" >> makefile.txt
 echo "prog: $(ls | grep *.c)" >> makefile.txt
 echo "	gcc -g -o prog $(ls | grep *c)" >> makefile.txt
